[not verified] Move fixtures to a separate file
