+++ projects/plugins/jetpack/_inc/client/recommendations/summary/test/component.js
@@ -11,6 +11,7 @@ import sinon from 'sinon';
 /**
  * Internal dependencies
  */
+import buildInitialState from './fixture';
 import { Summary as SummaryFeature } from '../index';
 import * as recommendationActions from 'state/recommendations/actions';
 
@@ -25,75 +26,6 @@ function render(
 	const Wrapper = ( { children } ) => <Provider store={ store }>{ children }</Provider>;
 
 	return rtlRender( ui, { wrapper: Wrapper, ...renderOptions } );
-}
-
-function rewindFixture( rewindStatus ) {
-	return {
-		data: {
-			status: rewindStatus,
-		},
-	};
-}
-
-function siteDataFixture( { productSlug }) {
-	return {
-		data: {
-			plan: {
-				product_slug: productSlug,
-			},
-		},
-	};
-}
-
-function upsellFixture( { hideUpsell } ) {
-	return {
-		product_id: 2101,
-		billing_timeframe: 'billed monthly',
-		cost_timeframe: 'per month',
-		cta_text: 'Learn more',
-		currency_code: 'USD',
-		description:
-			'Never lose a word, image, page, or time worrying about your site with automated off-site backups and one-click restores.',
-		features: [
-			'Automated daily off-site backups',
-			'One-click restores',
-			'Unlimited secure storage',
-		],
-		header: 'Recommended premium product',
-		hide_upsell: hideUpsell,
-		price: 9.95,
-		title: 'Backup Daily',
-	};
-}
-
-function buildInitialState( { hideUpsell = false, productSlug, rewindStatus = { state: 'unavailable' } } = {} ) {
-	return {
-		jetpack: {
-			initialState: {
-				userData: {
-					currentUser: 100,
-				},
-			},
-			connection: {
-				user: {
-					currentUser: {
-						isConnected: true,
-					},
-				},
-			},
-			pluginsData: {
-				items: [],
-			},
-			recommendations: {
-				upsell: upsellFixture( { hideUpsell } ),
-			},
-			rewind: rewindFixture( rewindStatus ),
-			settings: {
-				items: [],
-			},
-			siteData: siteDataFixture( { productSlug } ),
-		},
-	};
 }
 
 describe( "Recommendations – Summary section", () => {
